chore: Restructure Spotify package
