+++ pkg/service/spotify/models.go
@@ -1,4 +1,4 @@
-package models
+package spotify
 
 import "time"
